Incomplete game code. Temp move onto evolution
This code is non-functional and will be fixed later. COVID-19 interfered
with progress. For now, I will start simple weight evolution, assuming
that each individual will play a tetris game (calling the game code).
Thus, evolution will assume that this code works. The game code will
take an individual’s genome of weights (for features of the current
gamestate) and use a programmed function to determine what moves to
make. These weights will be evolved over generations. Eventually, the
“features” to consider, aka strategies, will be themselves evolved
(such that given basic info about the state, programs will evolve what
and how to consider the state for gameplay)

Each individual will play a game of tetris, choosing moves based on
features of the game state, each of which will be considered more or
less relative to other features. These weights, that individuals use to
decide moves when playing, will be evolved in the first simple
Evolutionary algorithm.